requirements.txt
pydantic>=2.0.0
openai>=1.0.0
pyyaml>=6.0.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
streamlit
fastapi
uvicorn
openenv-core>=0.2.0
uv